edita nome da lista

+++ game-slot/frontend/src/pages/Lists/ViewList/index.tsx
@@ -4,14 +4,11 @@ import { useParams, Redirect, useHistory } from 'react-router-dom'
 import { useDispatch } from 'react-redux'
 import classnames from 'classnames'
 import GamesGrid from '../../../components/GamesGrid'
-import {
-  deleteGameList as deleteGameListAction,
-  // deleteGameListItem,
-} from '../../../store/lists/actions'
+import { deleteGameList as deleteGameListAction } from '../../../store/lists/actions'
 import {
   deleteGameList,
-  // removeGameFromList,
   getOneGameList,
+  updateGameList,
 } from '../../../services/gameLists.service'
 import { useCurrentUser } from '../../../contexts/UserContext'
 import { useAuth0 } from '@auth0/auth0-react'
@@ -24,11 +21,12 @@ const ViewList: React.FC = () => {
   const [blocked, setBlocked] = useState<boolean>(false)
   const [loading, setLoading] = useState<boolean>(true)
   const { id } = useParams<Params>()
-  const [list, setList] = useState<GameList | null>()
+  const [list, setList] = useState<GameList | null>(null)
   const { user } = useCurrentUser()
-  const history = useHistory()
   const dispatch = useDispatch()
   const { getAccessTokenSilently } = useAuth0()
+  const [editing, setEditing] = useState<boolean>(false)
+  const history = useHistory()
 
   useEffect(() => {
     getOneGameList(id).then((listResponse) => {
@@ -37,16 +35,21 @@ const ViewList: React.FC = () => {
     })
   }, [id])
 
-  const onAddGames = () => {
-    history.push('/search')
+  const onSave = async () => {
+    const mayUpdate = confirm('Are you sure you want to edit this list?')
+    if (mayUpdate && user && list) {
+      const token = await getAccessTokenSilently()
+      await updateGameList(user._id, list, token)
+    }
   }
 
   const onDeleteList = async () => {
     const mayDelete = confirm('Are you sure you want to delete this list?')
-    if (mayDelete) {
+    if (mayDelete && user) {
       setBlocked(true)
       const token = await getAccessTokenSilently()
-      const success = await deleteGameList(id, token)
+      const success = await deleteGameList(id, user._id, token)
+      console.log(success)
       if (success) {
         dispatch(deleteGameListAction(id))
         setList(null)
@@ -73,15 +76,32 @@ const ViewList: React.FC = () => {
     history.push(`/game/${id}`)
   }
 
+  const onChangeListName = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setList(list ? { ...list, name: e.target.value } : null)
+  }
+
   if (loading) return <div />
 
   if (!list) return <Redirect to="/" />
 
   return (
     <section className="px-2 md:px-5">
-      <h1 className="tracking-widest text-white font-bold text-2xl uppercase py-3">
-        {list.name}
-      </h1>
+      <input
+        style={{
+          color: 'white',
+          fontWeight: 600,
+          backgroundColor: editing ? '#36383B' : '#1D1F22',
+          textTransform: 'uppercase',
+          fontSize: 24,
+          paddingTop: 12,
+          paddingBottom: 12,
+          paddingLeft: 16,
+          letterSpacing: 1.6,
+        }}
+        value={list.name}
+        disabled={!editing}
+        onChange={onChangeListName}
+      />
       <div
         className={classnames(
           'flex flex-col w-full xl:flex-row flex-wrap mt-2',
@@ -93,7 +113,11 @@ const ViewList: React.FC = () => {
         <button
           disabled={blocked}
           type="button"
-          onClick={onAddGames}
+          onClick={() => {
+            setEditing(!editing)
+            editing && onSave()
+          }}
+          style={{ backgroundColor: editing ? 'green' : 'blue' }}
           className={classnames(
             'bg-blue-600 text-white p-5 w-full text-center block md:inline xl:w-1/6 uppercase font-bold text-lg',
             {
@@ -101,7 +125,7 @@ const ViewList: React.FC = () => {
             }
           )}
         >
-          edit
+          {editing ? 'Save' : 'Edit'}
         </button>
         <button
           disabled={blocked}
@@ -114,7 +138,7 @@ const ViewList: React.FC = () => {
             }
           )}
         >
-          delete
+          Delete
         </button>
       </div>
       <h1 className="text-white mt-10 uppercase">games on this list</h1>
